+++ src/lib/components/sponsors/SponsorImage.svelte
@@ -10,7 +10,7 @@
 	const size = {
 		Tier1: 'h-16 menu-lg:h-24',
 		Tier2: 'h-20 menu-lg:h-26',
-		Tier3: 'h-36 menu-lg:h-44',
+		Tier3: 'h-28 menu-md:h-36 menu-lg:h-44',
 		Food: 'h-10 menu-lg:h-16'
 	}[tier];
 </script>
